@@ -7,17 +7,17 @@
   "dependencies": {
     "@mdx-js/loader": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
-    "@next/mdx": "^10.1.3",
+    "@next/mdx": "^12.0.7",
     "@react-spring/demo": "*",
     "@react-spring/parallax": "*",
     "clsx": "^1.1.1",
     "lodash-move": "^1.1.1",
     "lodash.clamp": "^4.0.3",
     "lodash.shuffle": "^4.2.0",
-    "lorem-ipsum": "^2.0.3",
-    "next": "^11.1.2",
-    "next-seo": "^4.23.0",
-    "next-transpile-modules": "^8.0.0",
+    "lorem-ipsum": "^2.0.4",
+    "next": "^12.0.7",
+    "next-seo": "^4.28.1",
+    "next-transpile-modules": "^9.0.0",
     "prism-react-renderer": "^1.2.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -28,12 +28,12 @@
     "react-spring": "*",
     "react-twitter-widgets": "^1.10.0",
     "react-use-gesture": "^9.1.3",
-    "react-use-measure": "^2.0.4",
+    "react-use-measure": "^2.1.1",
     "react-visibility-sensor": "^5.1.1",
-    "remark-autolink-headings": "^6.0.1",
-    "remark-slug": "^6.0.0",
+    "remark-autolink-headings": "^7.0.1",
+    "remark-slug": "^7.0.1",
     "resize-observer-polyfill": "^1.5.1",
-    "styled-components": "^5.2.3",
+    "styled-components": "^5.3.3",
     "vec-la": "^1.5.0"
   },
   "scripts": {
