@@ -1,8 +1,11 @@
-const withTM = require('next-transpile-modules')(['@react-spring/demo'])
-const slugs = require('remark-slug')
-const autoLinkHeadings = require('remark-autolink-headings')
+import transpile from 'next-transpile-modules'
+import slugs from 'remark-slug'
+import autoLinkHeadings from 'remark-autolink-headings'
+import mdx from '@next/mdx'
 
-const withMDX = require('@next/mdx')({
+const withTM = transpile(['@react-spring/demo'])
+
+const withMDX = mdx({
   extension: /\.mdx$/,
   options: {
     remarkPlugins: [
@@ -18,10 +21,11 @@ const withMDX = require('@next/mdx')({
   },
 })
 
-module.exports = withTM(
+export default withTM(
   withMDX({
     pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'md', 'mdx'],
     experimental: {
+      esmExternals: true,
       externalDir: true,
     },
     typescript: {
